Move InertExecutor into submodule of test_support and add basic coverage test.
